+++ RetroFE/Source/Graphics/Page.h
@@ -40,8 +40,9 @@ public:
     {
         ScrollDirectionForward,
         ScrollDirectionBack,
-        ScrollDirectionIdle
-
+        ScrollDirectionIdle,
+        ScrollDirectionPlaylistForward,
+        ScrollDirectionPlaylistBack,
     };
 
     Page(Configuration &c, int layoutWidth, int layoutHeight );
@@ -122,6 +123,7 @@ public:
     std::string controlsType() const;
     void setControlsType(const std::string& type);
     void  menuScroll();
+    void playlistScroll();
     void  highlightEnter();
     void  highlightExit();
     void  playlistEnter();
@@ -154,7 +156,7 @@ public:
     void  resetScrollPeriod() const;
     void  updateScrollPeriod() const;
     bool  isMenuFastScrolling() const;
-    void  scroll(bool forward);
+    void  scroll(bool forward, bool playlist);
     bool  hasSubs();
     int   getLayoutWidth(int layout);
     int   getLayoutHeight(int layout);